tweaks based on reviews
